fix rubocop linter errors
