--- Student.rb
@@ -1,14 +1,0 @@
-require_relative 'Person.rb'
-
-class Student < Person
-
-  def initialize(age, name = 'Unknown', parent_permission = true, classroom = 'None')
-    super(age, name, parent_permission)
-    @classroom = classroom
-  end
-
-  def play_hooky
-    "¯\(ツ)/¯"
-  end
-
-end
