+++ student_class.rb
@@ -1,0 +1,12 @@
+require_relative 'Person'
+
+class Student < Person
+  def initialize(age, name: 'Unknown', parent_permission: true, classroom: 'None')
+    super(age, name, parent_permission)
+    @classroom = classroom
+  end
+
+  def play_hooky
+    '¯(ツ)/¯'
+  end
+end
